requirements.txt
langchain==0.1.1
langchain-community==0.0.13
langchain-openai==0.0.2.post1
python-dotenv==1.0.0